added a repr function (probably have to rework to summary function)
