+++ pypesto/visualize/waterfall.py
@@ -4,8 +4,10 @@ import matplotlib.pyplot as plt
 import numpy as np
 from matplotlib.ticker import MaxNLocator
 
+from pypesto.util import delete_nan_inf
+
 from ..result import Result
-from .clust_color import RGBA, assign_colors, delete_nan_inf
+from .clust_color import RGBA, assign_colors
 from .misc import (
     process_offset_y,
     process_result_list,
